requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
APScheduler==3.6.3
asgiref==3.11.1
attrs==25.4.0
cachetools==4.2.2
certifi==2026.1.4
Django==6.0.2
frozenlist==1.8.0
idna==3.11
multidict==6.7.1
pillow==12.1.1
propcache==0.4.1
PySocks==1.7.1
python-dotenv==1.2.1
pytz==2025.2
setuptools==82.0.0
six==1.17.0
sqlparse==0.5.5
telepot==12.7
tornado==6.5.4
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
yarl==1.22.0
